+++ src/data_integrate/src/features/rel_point.cpp
@@ -1,4 +1,4 @@
-#include "../../include/features/rel_point.h"
+#include "data_integrate/features/rel_point.h"
 
 #include <angles/angles.h>
 #include <ros/ros.h>
