+++ src/data_integrate/src/features/visible_feature_manager.cpp
@@ -1,4 +1,4 @@
-#include "../include/features/visible_feature_manager.h"
+#include "data_integrate/features/visible_feature_manager.h"
 
 #include <ros/ros.h>
 #include <cmath>
