+++ src/data_integrate/src/tasks/blue_ball_capture_task.cpp
@@ -1,4 +1,4 @@
-#include "../../include/tasks/blue_ball_capture_task.h"
+#include "data_integrate/tasks/blue_ball_capture_task.h"
 
 #include <ros/ros.h>
