+++ src/data_integrate/src/test_simple_wheel_controller_node.cpp
@@ -6,7 +6,7 @@
 #include <ros/ros.h>
 #include <std_msgs/Float64.h>
 
-#include "../include/simple_wheel_controller.h"
+#include "data_integrate/simple_wheel_controller.h"
 
 /**
  * 현재 로봇이 실행하고 있는 운행 상태를 나타냅니다.
